Mirror mode support (sync HMD viewpoint with regular 3D view)
Adds option "Mirror HMD View" to the HMD panel in the 3D view.

In this implementation we update/redraw the mirrored 3D views with the
HMD view, which can actually cause some extra slowdown (drawing the
viewport at least once more). We decided that this would be the best
solution for now, as the other option would be opening a separate window
showing the mirrored result, which is a bit too much against the
non-overlapping paradigma IMHO.
We could also draw the HMD view into some buffer and reuse this for
mirrored displays, but that would mess up projection matrices.
In Blender 2.8 we'll probably be able to use multi-threaded drawing
anyway, for which this would be the perfect use case.

3D View regions mirroring the HMD view are locked, for quadview we only
use the non-locked region for mirroring.