Add/use template for browsing layout-types of a workspace/window combination
Tried adding a more generic template for this first, but this is really
such a special case that it would need to have quite a bunch of
exceptions.
Basically we're showing the active layer-instance of the workspace in
the text button and within the menu, the layout-type list of the
workspace and their previews from the screen wrapped by the
layout-instance :S

Also made renaming work properly.