Store active layout of each workspace per window
Allows users having a different active window stored even in inactive
workspaces per window. So that when activating a workspace the layout
is activated that was active the last time the workspace was active in
this window.
This is basically a per window, per workspace storage, but it's probably
how users would expect things to work.

Also, getting file read/write to work correctly was a challenge once
again, had to use global oldnew-map for pointer lookup on file read.

And another 'also': had to change order in which IDs are freed, so that
workspaces are freed after window-manager. Needed so that we can free
assignments/relations correctly when closing windows.