Allow choosing workspace from user config when adding workspace
When clicking the '+' icon to add a new workspace, a menu is spawned
now, showing the option 'Duplicate Current', but also a list of all
workspaces stored in a 'workflow.blend' as part of the user
configuration (next to startup.blend and userpref.blend)
Note that this workflow.blend has to be stored manually into the user
config folder right now. Of course it's not supposed to stay like this,
it should all be handled through simple operators for adding and removing
workspaces from the user configuration. We need to find out how that
would work UI wise though (tm).

When choosing a workspace from the menu, Blender does nothing but
appending the workspace from the workflow.blend file.

This design allows having a custom workspace setup stored as part of
the user configuration, it's however still possible to append workspaces
from regular .blend files.
This is useful e.g. if users want to open their workspaces without
overriding the workflow.blend when using the machine of someone else.