Rename editors/workspace back to editors/screen
We'll keep the name 'screen' to describe everything visible within the
window, including the global bars (top bar, status bar) which are not
part of bScreen struct (which describes the customizable screen-layout).
It's quite confusing to have the term 'screen' used for bScreen
operations too, but plan is to get rid of bScreen and move its data to
WorkSpace and WorkSpaceLayout (could be called ScreenLayout then).