Allow storing workspaces as part of user config (UI preliminary)
Adds a button in the User Preference Interface tab to save all
workspaces of the current file into [user config
path]/config/worflow.blend.
When clicking the button to add a new workspace, the menu will list
these then (and of course the option to duplicate current workspace).

Note that the operator currently overrides all workspaces in the config
file. There should be ways to add and remove workspaces from it instead,
but we need to figure out a good UI design for how to do that still.