Remove check on back of camera
This check is getting weird result and create a segfault in the cache due changes in the index of the stroke depending of point of view.

Need to have a better implementation. We keep disabled while we found a better solution that not change the index of the strokes (maybe separate this check and add to cache_idx anyway)