Refactor workspace data accessing
Following Campbell's and Bastien's suggestion to avoid the
all-or-nothing approach of using private headers.
Instead of using private headers we do some trickery to mark struct
members or whole structs as private.
This way we don't have to go overkill with accessors/mutators but keep
them where it makes sense.

For now this adds a macro based on DNA_DEPRECATED, just to mark certain
struct members as workspace-private (throws 'deprecated' warning if used
outside of workspace.c -- this doesn't work on MSVC though, not sure on
Clang).

Also, this adds a macro to mark struct members as workspace and file
read/write private, not sure if this is really worth it. Just thought it
might be worth considering, since some data is really workspace internal
but needs to be exposed for file read/write.