Implement lock axis in Sculpt grab brush
This is the first implementation to lock the axis while sculpt. In some extreme angles the result is not perfect, so maybe need a review.

We keep in this way while we found a better approach, maybe using the reprojection of the point as it is done when draw new strokes could work better.