Minimize Alpha stroke issue
Here we need is OIT (Order Independant Transparency) which is still an open issue in the realtime field (and also for Eevee). The best OIT method I know of is somewhat complex and not very precise (still not 1:1 with the ordered version). It involves more framebuffers and performance cost.

I may need to implement that in the draw manager in the future so that all engines can use it.

Thanks to Clement Foucault for the support provided.