Fix color distorsion while drawing
Need to use the mix_pass in order to get the same color while drawing. If the pass is not mixed with the same pass than strokes, there is a change in the color between drawing stroke and final stroke due the process to mix.

This is slower, but needed and while drawing the time delay is negligible