Fix T52100: Buffer stroke is not visible if paint over fill
If paint  a stroke over a fill area, the drawing stroke was not visible while drawing and only when finish the stroke.

The drawing pass was mixing the with mix pass and now reuse the stroke mix. This solve the problem and increase the drawing speed because only one mix pass is necessary