More work on modifiers
Complete rework, now the modifier data is saved in a cache and is not recalculated for each redrawing. The previous version had the problem the array strokes suffered of effect acumulation.

The array modifier is pending to add the rotation and scale parameters. Now only appear in UI panel but don't work.