Workspace: SETLOOPER should get render layer from workspace
Right now we are still passing NULL instead of the workspace in most of
those cases. That shouldn't matter much since this will simply bring
back the previous behaviour.

Also, most of those calls (gpu_material.c, view3d_draw_legacy.c,
render_internal.c) are from deprecated areas of Blender anyways.

The areas that will need to be addressed are:
* physics (collision.c)
* particles (pointcache.c)
* sound (sound.c)

And of course:
* game engine (BL_BlenderDataCoversion.cpp)

It's not clear to me, however, why some of those areas are using the
SET_LOOPER in the first place. SET_LOOPER acts on top of bases so we
want to iterate over the SceneLayer objects, not all of scene ones.

However in some cases it seems that the original design was expecting to
loop over all the objects (+ the set objects).