Properties Editor: Add Workspace and organize context path
We now have the properties of:

Scene, Scene > Layer, Scene > World, Workspace

[Scene | Workspace] > Render Layer > Object
[Scene | Workspace] > Render Layer > Object > Data
(...)

We use Scene or Workspace depending on workspace.use_scene_settings.

Also, we no longer expose the object in the collection context.
Since this should go away (and at the very least we should get the mode
from the workspace).