New Information Panel
This panel shows information about number of layers, frames, strokes, points and palettes used.

Notes: The number of frames in the internal data, not the number of different keyframes number.

The panel is not updated automatically to avoid performace issues. Close and reopen panel to refresh. Maybe, this could be changed, but need to be checked before.