Code Cleanup / Compile Fix
Further cleanups for the 2D monkey primitive - this time, to move
all the data arrays out of the single massive function, and out to file level.

Hopefully this solves the compile problems with GCC.

If nothing else, the file now compiles much faster here on MSVC
(i.e. previously, it was taking ~10-20 seconds to compile)