Tweak behaviour of adding a new GP Palette (to match other parts of Blender)
When creating a new Palette with an existing Palette assigned to a slot,
a copy of the old Palette will be created (i.e. the new copy will have all
the same colours), instead of creating a blank palette.

This makes it easier to create different versions of the same palette.