Fix bug of memory corruption
The strokes were transparent and zdepth was doing weird things when using 2 GP objects. This error was introduced by commit: 27/09/2017 08:34 5c8e1dfc2851a539abaa988762f1c590038e51bb

The problem was the field xray_mode. If move below in the struct, the value is corrupted and draw manager does not work fine..