GP Modifiers: Replace defunct "Duplication" modifier with "Build" modifier
It doesn't actually do anything now, and most of its properties are
undefined as well, but we don't want to break any existing files now
(as renumbering the other modifiers would cause issues)

Note: BKE_gpencil_reset_modifiers() has been removed. If the array modifier
needs this, we will have to find another solution to this.