Smooth new points while drawing
A new smooth algorithm is used for each new buffer point using the last two points to calculate the estimated position and average this point with the current mouse position.

This smooth reduce the jitter of the stroke and allows to use lower post smooth values to reduce stroke change after finish stroke.

Now the factor is hardcoded, but after some more testing maybe we need add a parameter by brush.