New parameter to adjust stabilization
Parameter to control the stabilization amount of the brush.

This stabilization works as a dynamic smooth using the previous points to calculate the estimated position and average the position to get smoother strokes.

This is not a replace of smooth factor that works after finish stroke. This new parameter is applyed while drawing before ending the stroke.