Cycles: Minor reworking of shader eval functions
Merges `shader_eval_surface` and `shader_eval_background` into the new
`shader_eval`. Also makes `kernel_shader_eval` of the split kernel
more generic. This will be used to split `svm_eval_nodes` out of
as many functions as possible so that there is only one call to this
function (or as close as we can get to that).