Remove the guide_object pointer from HairSystem.
Scalp mesh now is passed explicitly to the hair API functions again.

The idea is that hair systems can be created in any object and use
another object as the scalp. Sharing the same hair pattern should be
implemented at a later time, pending a good design for assigning
follicles to specific hair systems.