Fix: Temporary fix for crashes when moving GP keyframes when toggling Onion Skinning
ANIM_list_elem_update() was not doing anything as ale->id was null,
and we were filtering the channels by "animdata" blocks (i.e. the GP
datablocks). While technically what we're doing here isn't totally
correct (I'll need to come up with a better fix tomorrow), this will
be a good quick-fix for the crashes they animators are currently getting.