Rename "Hair" modifier to "Fur" modifier.
This will be a simple testbed for generating hair geometry.

The plan here is to use different modifiers for different editing methods
for hair. Each of these will use a HairSystem instance, but vary
in the editable data used to generate guide curves. The hair system
underneath is agnostic to the editing method used.