Fix: Crash when moving GP frame so that it replaces the one before it
This crash only occurred in a particular file, when onion skinning was
enabled. For some reason, the fact that this function was clearing the
layer's actframe (when the frame being deleted wasn't currently active)
was causing problems. I'm guessing that the GP draw engine code currently
assumes that gpl->actframe will always be set/valid?