WIP: New paper function
This function allows to cover full viewport with a predefined color. This allows to hide the geometry in complex scenes and make the drawing area cleaner.

This function was requested by artist after working in Hero open movie using complex scenes with a lot of geometry "behind" of the grease pencil drawings.