Avoid some operators in multiframe edit mode
As initial idea is better to keep limited to single edit mode, but maybe in the future we can enable it.

The reason to limit them is because in some situations the artist can get confused with the edition he did if multiframe is activated and could consider the operation as a bug when really was working as expected.