Fix linking from temp filebrowser.
We should only set mouse coordinates when we are actually invoked to
link something, not when we are invoked to show filebrowser.

Else (filebrowser invoke case) we have a mouse coordinates sticking
around, which we try to use later in wrong context.

For some reasons, in master that was not an issue, but in 2.8
ED_view3d_give_base_under_cursor() will crash when context is wrong. :/