Make Amber a real add-on.
Well, real add-on... since we want to keep it in main repo for now, it
still needs some hacks to 'behave' like a regular add-on. :/

Anyway, it's now removed from startup files, and will live in modules
until we go to master itself.

Also, make it enabled by default in factory startup.