New VFX light modifier
This modifier simulates a point light over the stroke. There is not shadows support and use its own internal lamp point. Only ambient and diffuse.

The use of this VFX is to get special effects. In the future, a real lighting system can be implemented using normal Blender lamps, but this is different of this modifier.