Fix problem with merge down
The problem was if the layer below had animation in frames not present in top layer. As the active frame of top layer was the previous keyframe, after merging, the draw of bottom layer replaced the draw of top layer and the result was that some drawings were missing.

The solution was to duplicate the active frame in top layer before merge the bottom layer.