New no fill strokes for fill boundaries
This special strokes can only be generated using a fill brush and pressing Alt key.

This type of strokes are required to draw fast boundary with the fill color for opened areas to use the automatic fill later.

The color of these strokes will be the fill color and not the stroke color.