Fix error in OpenGL render
As F12 and OpenGL are detected as render mode, the pixel size was not correct.

I have added a check to verify if it is a real F12 render or it's a OpenGL.

Maybe this check must be removed if the OpenGL render operator is deprecated by a new system, but we need to keep the verification until them.