WIP: Improved use of UV along the path and speedup
Now the texture is calculated using the pixelsize along the stroke.

Added a precalculation of fill and Uvs caches in order to avoid recalculation on frame change. This makes first play a little more slow (few miliseconds), but the next time the frame is used, all cache data is precalculated and the overall FPS is better when use fill strokes.