Fast mouse event capture afected by zoom factor
Before, the number of points was not equal at different zoom levels.

To get this work, the pixel size is calculated and the distance is relative to 30% of the brush thickness multiplied by the number of samples.

The number of samples has been reduced too, because the new method doesn't need big values.